+++ Sources/KeyboardKit/Views/System/SystemKeyboardActionButton.swift
@@ -13,16 +13,17 @@ import SwiftUI
  keyboard. It wraps a `SystemKeyboardButton` view and uses a
  `SystemKeyboardActionButtonContent` to resolves its content.
  
+ Note that this view only mimics the look of a system button
+ for the provided `action` and `appearance` and does not add
+ any gestures or button actions to itself. When you do apply
+ gestures, make sure to change `isPressed` accordingly. This
+ is easiest to do by creating your own wrapper view.
+ 
  You can use the `contentConfig` to customize the content in
  the button to fit your needs.
  
- The view uses the provided `context` and `actionHandler` to
- perform actions when the button is tapped, which means that
- you can create this button anywhere and it will trigger the
- correct gesture actions for tap, long press, drag etc.
- 
  `SystemKeyboardButton` can be used to create a basic button
- that just looks correct, but doesn't do anything.
+ that have a fixed content and style.
  */
 public struct SystemKeyboardActionButton<Content: View>: View {
     
@@ -31,57 +32,37 @@ public struct SystemKeyboardActionButton<Content: View>: View {
      
      - Parameters:
        - action: The keyboard action to apply.
-       - actionHandler: The keyboard action handler to use.
        - appearance: The keyboard appearance to use.
+       - isPressed: Whether or not the button is pressed.
        - contentConfig: A content configuration block that can be used to modify the button content before applying a style and gestures to it.
      */
     public init(
         action: KeyboardAction,
-        actionHandler: KeyboardActionHandler,
         appearance: KeyboardAppearance,
+        isPressed: Binding<Bool>,
         contentConfig: @escaping ContentConfig) {
         self.action = action
-        self.actionHandler = actionHandler
         self.appearance = appearance
+        self._isPressed = isPressed
         self.contentConfig = contentConfig
     }
     
     private let action: KeyboardAction
-    private let actionHandler: KeyboardActionHandler
     private let appearance: KeyboardAppearance
     private let contentConfig: ContentConfig
+
+    @Binding private var isPressed: Bool
     
     /**
      This typealias represents the configuration action that
      is used to customize the content of the button.
      */
     public typealias ContentConfig = (SystemKeyboardActionButtonContent) -> Content
-    
-    @State private var isPressed = false
-    
-    @EnvironmentObject private var context: KeyboardContext
-    
+        
     public var body: some View {
-        button.keyboardGestures(
-            for: action,
-            context: context,
-            actionHandler: actionHandler,
-            isPressed: $isPressed)
-    }
-}
-
-private extension SystemKeyboardActionButton {
-    
-    var button: some View {
         SystemKeyboardButton(
             content: buttonContent,
             style: buttonStyle)
-    }
-    
-    var buttonStyle: SystemKeyboardButtonStyle {
-        appearance.systemKeyboardButtonStyle(
-            for: action,
-            isPressed: isPressed)
     }
 }
 
@@ -93,18 +74,17 @@ public extension SystemKeyboardActionButton where Content == SystemKeyboardActio
      
      - Parameters:
        - action: The keyboard action to apply.
-       - actionHandler: The keyboard action handler to use.
        - appearance: The keyboard appearance to use.
-       - contentConfig: A content configuration block, to adjust the content before presenting it.
+       - isPressed: Whether or not the button is pressed.
      */
     init(
         action: KeyboardAction,
-        actionHandler: KeyboardActionHandler,
-        appearance: KeyboardAppearance) {
+        appearance: KeyboardAppearance,
+        isPressed: Binding<Bool>) {
         self.init(
             action: action,
-            actionHandler: actionHandler,
             appearance: appearance,
+            isPressed: isPressed,
             contentConfig: { $0 })
     }
 }
@@ -118,30 +98,60 @@ private extension SystemKeyboardActionButton {
                 appearance: appearance)
         )
     }
+    
+    var buttonStyle: SystemKeyboardButtonStyle {
+        appearance.systemKeyboardButtonStyle(
+            for: action,
+            isPressed: isPressed)
+    }
 }
 
 struct SystemKeyboardActionButton_Previews: PreviewProvider {
     
-    static func button(for action: KeyboardAction) -> some View {
-        SystemKeyboardActionButton(
-            action: action,
-            actionHandler: .preview,
-            appearance: PreviewKeyboardAppearance()) {
-                $0.frame(width: 60, height: 60)
-            }
+    private struct Preview: View {
+        
+        let action: KeyboardAction
+        
+        @State private var isPressed = false
+        
+        var body: some View {
+            SystemKeyboardActionButton(
+                action: action,
+                appearance: .preview,
+                isPressed: $isPressed) {
+                $0.frame(width: 80, height: 80)
+                }.onPressGesture {
+                    isPressed = true
+                } onReleased: {
+                    isPressed = false
+                }
+        }
     }
     
     static var previews: some View {
         VStack {
-            button(for: .backspace)
-            button(for: .space)
-            button(for: .character("a"))
-            button(for: .character("A"))
+            Preview(action: .backspace)
+            Preview(action: .nextKeyboard)
+            Preview(action: .character("a"))
+            Preview(action: .character("A"))
         }
         .padding()
         .background(Color.gray)
         .cornerRadius(10)
         .environment(\.sizeCategory, .extraExtraLarge)
         .keyboardPreview()
+    }
+}
+
+private extension View {
+    
+    func onPressGesture(
+        onPressed: @escaping () -> Void,
+        onReleased: @escaping () -> Void) -> some View {
+        self.gesture(
+            DragGesture(minimumDistance: 0)
+                .onChanged { _ in onPressed() }
+                .onEnded { _ in onReleased() }
+        )
     }
 }